bar101-storyline-gen/pyproject.toml
[project]
name = "bar101-storyline-gen"
version = "0.1.0"
description = ""
authors = [
    {name = "Krzysztof Jamroz",email = "kjamroz83@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai (>=1.69.0,<2.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "rich (>=13.9.4,<14.0.0)",
    "questionary (>=2.1.0,<3.0.0)"
]

[tool.poetry]
packages = [{include = "bar101_storyline", from = "src"}]


[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
isort = "^6.0.1"
pytest = "^8.3.5"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
